# SOMS/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Cloud Platform API Server",
        "title": "SOMS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "117.16.136.172:3000",
    "basePath": "/",
    "paths": {
        "/action/approve/{id}": {
            "post": {
                "description": "VM 생성을 승인합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vm"
                ],
                "summary": "VM 생성 승인",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VM uuid",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "승인대상 사용자 UUID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vm.ApproveVMRequestBody"
                        }
                    },
                    {
                        "type": "string",
                        "description": "UUID",
                        "name": "X-UUID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/action/hardreboot/{id}": {
            "post": {
                "description": "VM을 하드 리부팅합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vm"
                ],
                "summary": "VM 하드 리부팅",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VM uuid",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID",
                        "name": "X-UUID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/action/poweroff/{id}": {
            "post": {
                "description": "VM을 종료합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vm"
                ],
                "summary": "VM 종료",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VM uuid",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID",
                        "name": "X-UUID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/action/poweron/{id}": {
            "post": {
                "description": "VM을 시작합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vm"
                ],
                "summary": "VM 시작",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VM uuid",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID",
                        "name": "X-UUID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/action/snapshot/{id}": {
            "post": {
                "description": "VM의 스냅샷을 생성합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vm"
                ],
                "summary": "VM 스냅샷 생성",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VM uuid",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "스냅샷 이름",
                        "name": "snapshotName",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "type": "string",
                        "description": "UUID",
                        "name": "X-UUID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/action/softreboot/{id}": {
            "post": {
                "description": "VM을 소프트 리부팅합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vm"
                ],
                "summary": "VM 소프트 리부팅",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VM uuid",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID",
                        "name": "X-UUID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/action/vnc/{id}": {
            "get": {
                "description": "VM의 VNC URL을 조회합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vm"
                ],
                "summary": "VM VNC URL 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VM uuid",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID",
                        "name": "X-UUID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
...
        "/resource/flavor": {
            "get": {
                "description": "VM의 플레이버 리스트를 조회합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vm"
                ],
                "summary": "VM 플레이버 리스트 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID",
                        "name": "X-UUID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/resource/image": {
            "get": {
                "description": "VM의 이미지 리스트를 조회합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vm"
                ],
                "summary": "VM 이미지 리스트 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID",
                        "name": "X-UUID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
...
        "/resource/securitygroup": {
            "get": {
                "description": "VM의 보안그룹 리스트를 조회합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vm"
                ],
                "summary": "VM 보안그룹 리스트 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID",
                        "name": "X-UUID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
...
        "/vmstat": {
            "get": {
                "description": "VM의 상태를 조회합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vm"
                ],
                "summary": "VM 상태 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID",
                        "name": "X-UUID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "response.CommonResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "vm.ApproveVMRequestBody": {
            "type": "object",
            "properties": {
                "approveUserUUID": {
                    "type": "string"
                }
            }
        },
...
    }
}